football-stats/requirements.txt
requests==2.31.0
python-dotenv==1.0.0
# Image processing and OCR
pillow==10.2.0
pytesseract==0.3.10
opencv-python==4.9.0.80
# Data analysis and visualization
pandas==2.2.0
matplotlib==3.8.2
tabulate==0.9.0